requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
supabase>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
fastapi-cors>=0.0.6

